backend/requirements.txt
accelerate==0.24.1
bitsandbytes==0.41.2.post2
fastapi==0.104.1
onnx==1.15.0
onnxruntime-gpu==1.16.2
optimum==1.14.1
python-decouple==3.8
transformers==4.30.0
uvicorn==0.24.0.post1
python-dotenv==1.0.0
scipy==1.11.3
torch